Corrections

+++ src/main/kotlin/trainer/model/Word.kt
@@ -5,8 +5,8 @@ data class Word(
     val translations: String,
     val type: String,
     var correctAnswersCount: Short = START_CORRECT_ANSWERS_COUNT,
-    var usageCount: Int = START_USAGE_COUNT
+    var usageCount: Short = START_USAGE_COUNT
 )
 
 const val START_CORRECT_ANSWERS_COUNT = 0.toShort()
-const val START_USAGE_COUNT = 0
+const val START_USAGE_COUNT = 0.toShort()
